requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
mysql-connector-python==8.0.33
python-dotenv==1.0.0
pandas==1.5.3
openpyxl==3.1.2